+++ Tests/PlayMode/ChangeColorObjectsTest.cs
@@ -3,13 +3,15 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.TestTools;
 using System.Linq;
+using System;
+using NUnit.Framework;
+
+using ReupVirtualTwin.models;
+using ReupVirtualTwin.modelInterfaces;
+using ReupVirtualTwin.helpers;
 using ReupVirtualTwin.managers;
 using ReupVirtualTwin.enums;
 using ReupVirtualTwin.managerInterfaces;
-using System;
-using ReupVirtualTwin.models;
-using ReupVirtualTwin.modelInterfaces;
-using NUnit.Framework;
 
 public class ChangeColorObjectsTest : MonoBehaviour
 {
